# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/astrologer/profile": {
      "get": {
        "operationId": "AstrologerProfileController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      },
      "patch": {
        "operationId": "AstrologerProfileController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAstrologerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
...
    "/api/v1/astrologer/availability": {
      "get": {
        "operationId": "AstrologerProfileController_getAvailability",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      },
      "patch": {
        "operationId": "AstrologerProfileController_updateAvailability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
    "/api/v1/astrologer/profile/working-hours": {
      "patch": {
        "operationId": "AstrologerProfileController_updateWorkingHours",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkingHoursDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
...
    "/api/v1/astrologer/profile/change-request": {
      "post": {
        "operationId": "AstrologerProfileController_requestProfileChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestProfileChangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
...
    "/api/v1/astrologer/earnings": {
      "get": {
        "operationId": "AstrologerProfileController_getEarnings",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
    "/api/v1/astrologer/stats": {
      "get": {
        "operationId": "AstrologerProfileController_getStats",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
    "/api/v1/astrologer/transactions": {
      "get": {
        "operationId": "TransactionController_getTransactions",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Transaction"
        ]
      }
    },
    "/api/v1/astrologer/transactions/stats": {
      "get": {
        "operationId": "AstrologerProfileController_getTransactionStats",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
    "/api/v1/astrologer/gifts/stats": {
      "get": {
        "operationId": "AstrologerProfileController_getGiftStats",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AstrologerProfile"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateAstrologerProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkingHoursDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAvailabilityDto": {
        "type": "object",
        "properties": {}
      },
      "RequestProfileChangeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}